Created new classes.
The Qr History manager as its name says, handles the history of the QR codes created by the passenger.
The redirect class works by redirecting the user based on it's email and password, with this it can identify if it's a passenger or a driver. Due the nature of our project - and since it's focused on the passenger- it works only with the passenger info at the moment.
Sign in as it name says handles the sign in of the user.